backend/pyproject.toml
[tool.poetry]
name = "stock-insight-backend"
version = "1.0.0"
description = "AI 기반 글로벌 주식 인사이트 플랫폼 - Backend"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
anthropic = "^0.40.0"
google-generativeai = "^0.8.0"
supabase = "^2.10.0"
yfinance = "^0.2.48"
feedparser = "^6.0.11"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
